requirements.txt
cliff
gitpython
github3.py
psutil
sh
jinja2
pyyaml